# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP API для управления отложенными уведомлениями (создание, проверка статуса и отмена рассылок).",
        "title": "Delayed Notifier API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/notify": {
            "post": {
                "description": "Принимает задачу для отложенной отправки уведомления.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Создать уведомление",
                "parameters": [
                    {
                        "description": "Данные уведомления",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreateNotificationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "UUID уведомления",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "Location": {
                                "type": "string",
                                "description": "URL созданного ресурса"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/notify/{id}": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Получить статус уведомления",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID уведомления",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Текущий статус",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Некорректный идентификатор",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Отменить уведомление",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID уведомления",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Некорректный идентификатор",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Channel": {
            "type": "integer",
            "enum": [
                0,
                1
            ],
            "x-enum-varnames": [
                "Telegram",
                "Email"
            ]
        },
        "requests.CreateNotificationRequest": {
            "type": "object",
            "properties": {
                "channel": {
                    "$ref": "#/definitions/models.Channel"
                },
                "message": {
                    "type": "string"
                },
                "recipient": {
                    "type": "string"
                },
                "scheduled_at": {
                    "type": "string"
                }
            }
        }
    }
}